Update weather.py
Updated the visibility to show either mi or km depending on units of measure setting.  Also updated to not run the conversion in the script for Open Meteo data source.  But will continue to show only in mi / km (Open Metro shows in ft/meters). Also update is non destructive where it will show N/A if the visibility is unavailable.